+++ packages/eslint-plugin-formatjs/rules/enforce-description.ts
@@ -1,9 +1,6 @@
 import {TSESTree} from '@typescript-eslint/utils'
-import {
-  RuleContext,
-  RuleModule,
-  RuleListener,
-} from '@typescript-eslint/utils/ts-eslint'
+import {RuleContext, RuleModule} from '@typescript-eslint/utils/ts-eslint'
+import {getParserServices} from '../context-compat'
 import {extractMessages, getSettings} from '../util'
 
 export enum Option {
@@ -46,7 +43,7 @@ function checkNode(
 
 export const name = 'enforce-description'
 
-export const rule: RuleModule<MessageIds, Options, RuleListener> = {
+export const rule: RuleModule<MessageIds, Options> = {
   meta: {
     type: 'problem',
     docs: {
@@ -73,9 +70,9 @@ export const rule: RuleModule<MessageIds, Options, RuleListener> = {
       checkNode(context, node)
 
     //@ts-expect-error defineTemplateBodyVisitor exists in Vue parser
-    if (context.parserServices.defineTemplateBodyVisitor) {
+    if (getParserServices(context).defineTemplateBodyVisitor) {
       //@ts-expect-error
-      return context.parserServices.defineTemplateBodyVisitor(
+      return getParserServices(context).defineTemplateBodyVisitor(
         {
           CallExpression: callExpressionVisitor,
         },