+++ packages/eslint-plugin-formatjs/rules/no-camel-case.ts
@@ -5,15 +5,11 @@ import {
   parse,
 } from '@formatjs/icu-messageformat-parser'
 import {TSESTree} from '@typescript-eslint/utils'
-import {
-  RuleContext,
-  RuleModule,
-  RuleListener,
-} from '@typescript-eslint/utils/ts-eslint'
+import {RuleContext, RuleModule} from '@typescript-eslint/utils/ts-eslint'
+import {getParserServices} from '../context-compat'
 import {extractMessages, getSettings} from '../util'
 
 type MessageIds = 'camelcase'
-type Options = [Element[]?]
 
 export const name = 'no-camel-case'
 
@@ -43,7 +39,7 @@ function verifyAst(ast: MessageFormatElement[]) {
 }
 
 function checkNode(
-  context: RuleContext<MessageIds, Options>,
+  context: RuleContext<MessageIds, unknown[]>,
   node: TSESTree.Node
 ) {
   const settings = getSettings(context)
@@ -72,7 +68,7 @@ function checkNode(
   }
 }
 
-export const rule: RuleModule<MessageIds, Options, RuleListener> = {
+export const rule: RuleModule<MessageIds> = {
   meta: {
     type: 'problem',
     docs: {
@@ -91,9 +87,9 @@ export const rule: RuleModule<MessageIds, Options, RuleListener> = {
       checkNode(context, node)
 
     //@ts-expect-error defineTemplateBodyVisitor exists in Vue parser
-    if (context.parserServices.defineTemplateBodyVisitor) {
+    if (getParserServices(context).defineTemplateBodyVisitor) {
       //@ts-expect-error
-      return context.parserServices.defineTemplateBodyVisitor(
+      return getParserServices(context).defineTemplateBodyVisitor(
         {
           CallExpression: callExpressionVisitor,
         },