+++ packages/eslint-plugin-formatjs/tests/enforce-placeholders.test.ts
@@ -1,27 +1,36 @@
-import {rule, name} from '../rules/enforce-placeholders'
+import {name, rule} from '../rules/enforce-placeholders'
+import {dynamicMessage, emptyFnCall, noMatch, spreadJsx} from './fixtures'
 import {ruleTester} from './util'
-import {dynamicMessage, noMatch, spreadJsx, emptyFnCall} from './fixtures'
 ruleTester.run(name, rule, {
   valid: [
-    `intl.formatMessage({
+    {
+      code: `intl.formatMessage({
       defaultMessage: '{count, plural, one {#} other {# more}}',
       description: 'asd'
   }, {count: 1})`,
-    `intl.formatMessage({
+    },
+    {
+      code: `intl.formatMessage({
     defaultMessage: '{count, plural, one {#} other {# more}}',
     description: 'asd'
   }, {'count': 1})`,
-    `import {FormattedMessage} from 'react-intl'
+    },
+    {
+      code: `import {FormattedMessage} from 'react-intl'
   const a = <FormattedMessage
   defaultMessage="{count, plural, one {#} other {# more}}"
   values={{ count: 1}} />
         `,
-    `import {FormattedMessage} from 'react-intl'
+    },
+    {
+      code: `import {FormattedMessage} from 'react-intl'
   const a = <FormattedMessage
   defaultMessage="{count, plural, one {#} other {# more}} {bar}"
   values={{ 'count': 1, bar: 2}} />
         `,
-    `import {defineMessages, _} from 'react-intl'
+    },
+    {
+      code: `import {defineMessages, _} from 'react-intl'
   defineMessages({
     foo: {
       defaultMessage: '{count, plural, one {#} other {# more}}',
@@ -33,7 +42,9 @@ ruleTester.run(name, rule, {
     description: 'asd'
   })
   `,
-    `
+    },
+    {
+      code: `
   intl.formatMessage({
     defaultMessage: '{count, plural, one {<a>#</a>} other {# more}}',
     description: 'asd'
@@ -42,7 +53,9 @@ ruleTester.run(name, rule, {
     a: (...chunks) => <a>{chunks}</a>
   })
   `,
-    `
+    },
+    {
+      code: `
   intl.formatMessage({
     defaultMessage: '{count, plural, one {<a>#</a>} other {# more}}',
     description: 'asd'
@@ -52,6 +65,7 @@ ruleTester.run(name, rule, {
     a: (...chunks) => <a>{chunks}</a>
   })
   `,
+    },
     dynamicMessage,
     noMatch,
     spreadJsx,