+++ packages/eslint-plugin-formatjs/tests/no-invalid-icu.test.ts
@@ -1,36 +1,49 @@
 import {name, rule} from '../rules/no-invalid-icu'
+import {dynamicMessage, emptyFnCall, noMatch, spreadJsx} from './fixtures'
 import {ruleTester} from './util'
-import {dynamicMessage, noMatch, spreadJsx, emptyFnCall} from './fixtures'
 
 ruleTester.run(name, rule, {
   valid: [
-    `intl.formatMessage({
+    {
+      code: `intl.formatMessage({
       defaultMessage: '{count, plural, one {#} other {# more}}',
       description: 'asd'
   }, {count: 1})`,
-    `this.intl.formatMessage({
+    },
+    {
+      code: `this.intl.formatMessage({
       defaultMessage: '{count, plural, one {#} other {# more}}',
       description: 'asd'
   }, {count: 1})`,
-    `IntlStore.intl.formatMessage({
+    },
+    {
+      code: `IntlStore.intl.formatMessage({
       defaultMessage: '{count, plural, one {#} other {# more}}',
       description: 'asd'
   }, {count: 1})`,
-    `intl.formatMessage({
+    },
+    {
+      code: `intl.formatMessage({
     defaultMessage: '{count, plural, one {#} other {# more}}',
     description: 'asd'
   }, {'count': 1})`,
-    `import {FormattedMessage} from 'react-intl'
+    },
+    {
+      code: `import {FormattedMessage} from 'react-intl'
   const a = <FormattedMessage
   defaultMessage="{count, plural, one {#} other {# more}}"
   values={{ count: 1}} />
         `,
-    `import {FormattedMessage} from 'react-intl'
+    },
+    {
+      code: `import {FormattedMessage} from 'react-intl'
   const a = <FormattedMessage
   defaultMessage="{count, plural, one {#} other {# more}} {bar}"
   values={{ 'count': 1, bar: 2}} />
         `,
-    `import {defineMessages, _} from 'react-intl'
+    },
+    {
+      code: `import {defineMessages, _} from 'react-intl'
   defineMessages({
     foo: {
       defaultMessage: '{count, plural, one {#} other {# more}}',
@@ -42,7 +55,9 @@ ruleTester.run(name, rule, {
     description: 'asd'
   })
   `,
-    `
+    },
+    {
+      code: `
   intl.formatMessage({
     defaultMessage: '{count, plural, one {<a>#</a>} other {# more}}',
     description: 'asd'
@@ -51,7 +66,9 @@ ruleTester.run(name, rule, {
     a: (...chunks) => <a>{chunks}</a>
   })
   `,
-    `
+    },
+    {
+      code: `
   intl.formatMessage({
     defaultMessage: '{count, plural, one {<a>#</a>} other {# more}}',
     description: 'asd'
@@ -61,6 +78,7 @@ ruleTester.run(name, rule, {
     a: (...chunks) => <a>{chunks}</a>
   })
   `,
+    },
     dynamicMessage,
     noMatch,
     spreadJsx,